# Library/Swagger/docs/swagger.yaml
basePath: /employee
definitions:
...
  Dto.LeaveBalanceAdjustmentDTO:
    properties:
      createdBy:
        type: string
      createdTime:
        type: string
      deleted:
        type: boolean
      employeeCode:
        maxLength: 50
        type: string
      employeeId:
        type: integer
      endDate:
        type: string
      id:
        type: integer
      leaveId:
        type: integer
      quantity:
        type: integer
      reason:
        type: string
      startDate:
        type: string
      tahun:
        type: integer
      type:
        type: string
    required:
    - employeeCode
    type: object
...
  Response.RespErrorStruct:
    properties:
      message:
        type: string
      result:
        additionalProperties:
          type: string
        example:
          key: value
          key2: value2
        type: object
      status:
        example: 400
        type: integer
      title:
        example: bad_request
        type: string
    type: object
  Response.RespResultStruct:
    properties:
      message:
        type: string
      result:
        additionalProperties:
          type: string
        example:
          key: value
          key2: value2
        type: object
      status:
        type: integer
      title:
        type: string
    type: object
info:
  contact: {}
  description: 'Example: Bearer abcdefghijklmnopqrstuvwxyz12345678901'
  title: Employee Service
  version: "1.0"
paths:
...
  /leave-balance:
    get:
      consumes:
      - application/json
      description: 'Sample Parameter: `?sortColumn=id&sortOrder=desc&pageSize=20?companyName=com&locationName=loc&departmentName=dept`'
      parameters:
      - description: sortColumn
        in: query
        name: sortColumn
        type: string
      - description: sortOrder
        in: query
        name: sortOrder
        type: string
      - description: pageSize
        in: query
        name: pageSize
        type: integer
      - description: pageNumber
        in: query
        name: pageNumber
        type: integer
      - description: companyName
        in: query
        name: companyName
        type: string
      - description: locationName
        in: query
        name: locationName
        type: string
      - description: departmentName
        in: query
        name: departmentName
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Response.RespResultStruct'
        "500":
          description: desc
          schema:
            $ref: '#/definitions/Response.RespErrorStruct'
      security:
      - Bearer: []
      summary: Show Leave Balance Data
      tags:
      - Leave Balance
    post:
      consumes:
      - application/json
      description: 'Sample Payload: `{"employeeId": 36,"employeeCode":"10.0001","startDate":"2022-02-10T07:48:48.243Z","endDate":"2022-02-10T07:48:48.243Z","type":"increase","quantity":
        5,"reason":"Potongan Cuti 15 Hari"}`'
      parameters:
      - description: Example Payload
        in: body
        name: '"payload"'
        required: true
        schema:
          $ref: '#/definitions/Dto.LeaveBalanceAdjustmentDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Response.RespResultStruct'
        "500":
          description: desc
          schema:
            $ref: '#/definitions/Response.RespErrorStruct'
      security:
      - Bearer: []
      summary: Create Leave Balance Adjusment
      tags:
      - Leave Balance
  /leave-balance/{employeeId}:
    get:
      consumes:
      - application/json
      description: 'Sample Parameter: `2`'
      parameters:
      - description: employeeId
        in: path
        name: employeeId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Response.RespResultStruct'
        "500":
          description: desc
          schema:
            $ref: '#/definitions/Response.RespErrorStruct'
      security:
      - Bearer: []
      summary: Show Leave Balance Detail Data
      tags:
      - Leave Balance
  /leave-balance/{employeeId}/{year}/history:
    get:
      consumes:
      - application/json
      description: 'Sample Parameter: `123/2021`'
      parameters:
      - description: employeeId
        in: path
        name: employeeId
        required: true
        type: integer
      - description: year
        in: path
        name: year
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Response.RespResultStruct'
        "500":
          description: desc
          schema:
            $ref: '#/definitions/Response.RespErrorStruct'
      security:
      - Bearer: []
      summary: Show Leave Balance Adjustment Detail
      tags:
      - Leave Balance
...
  /leave-balance/reset-leave-balance:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Response.RespResultStruct'
        "500":
          description: desc
          schema:
            $ref: '#/definitions/Response.RespErrorStruct'
      security:
      - Bearer: []
      summary: Reset Leave Balance
      tags:
      - Leave Balance
      - Leave Balance
...
securityDefinitions:
  Bearer:
    in: Header
    name: Authorization
    type: apiKey
swagger: "2.0"